Deployment to Rinkeby TestNet
Created the deploy.js file to deploy the Inbox Smart Contract.  Created a Secrets.js file which holds the Infura Key and the Mnemonic to the accounts. Added the secrets to gitignore. Also updated the Deploy.js file to call the read the message that was set from the Rinkeyby deployed SC (from the blockchain).